==> notebooks/1_initial_inspection.ipynb <==
import pandas as pd
import numpy as np

df = pd.read_csv("/Users/ananyaaggarwal/Desktop/Movie Prediction Project/data/tmdb_5000_movies.csv")

df.head()

df.tail()

df.describe()

df.info()

df.isnull().sum()
