==> notebooks/2_data_preprocessing.ipynb <==
import pandas as pd
import numpy as np
import ast

movie_df = pd.read_csv("/Users/ananyaaggarwal/Desktop/Movie Prediction Project/data/tmdb_5000_movies.csv")

credits_df = pd.read_csv("/Users/ananyaaggarwal/Desktop/Movie Prediction Project/data/tmdb_5000_credits.csv")

movie_df.rename(columns= {"id" : "movie_id"}, inplace=True)

merge_df = pd.merge(movie_df, credits_df, on = "movie_id")

merge_df.head(2)

merge_df.drop(["homepage", "keywords", "overview", "spoken_languages", "tagline", "title_x", "title_y"], inplace=True)

merge_df.drop("homepage" , axis = 1, inplace=True)

merge_df.drop("keywords" , axis = 1, inplace=True)

merge_df.drop("overview" , axis = 1, inplace=True)

merge_df.drop("spoken_languages" , axis = 1, inplace=True)

merge_df.drop("tagline" , axis = 1, inplace=True)

merge_df.drop("title_x" , axis = 1, inplace=True)

merge_df.drop("title_y" , axis = 1, inplace=True)

merge_df.head(2)

merge_df.dropna(subset = ["budget","revenue"], inplace=True)


def extract_top_cast(cast_list):
    if isinstance(cast_list, list):
        return [i.get('name', '') for i in cast_list[:3]]  # Extract top 3 names
    return []  # Return empty list for invalid data

merge_df['cast'].apply(extract_top_cast)


merge_df.head(2)

print(type(merge_df['cast'].iloc[0]))




def extract_top_cast(cast_data):
    if isinstance(cast_data, str):  # If it's a string, convert it
        try:
            cast_data = ast.literal_eval(cast_data)  # Convert to list
        except (ValueError, SyntaxError):
            return []  # Return empty if conversion fails

    if isinstance(cast_data, list):  # Ensure it's now a list
        return [i.get('name', '') for i in cast_data[:3]]  # Extract top 3 names

    return []  # Return empty if not a valid list


merge_df['cast'] = merge_df['cast'].apply(extract_top_cast)

merge_df.head(2)

merge_df

print(type(merge_df['crew'].iloc[0]))

def extract_director(crew_data):
    if isinstance(crew_data, str):
        try:
            crew_data = ast.literal_eval(crew_data)
        except(ValueError, SyntaxError):
            return ""
        
    if isinstance(crew_data, list):
        for i in crew_data:
            if i.get("job") == "Director":
                return i.get('name', '')
    return ""

merge_df["crew"].apply(extract_director)

merge_df.head(2)

merge_df['director'] = merge_df['crew'].apply(extract_director)

merge_df.head(2)

merge_df.drop("crew", axis = 1,inplace=True)

merge_df.head(2)

merge_df.drop("original_language" ,axis=1, inplace=True)

merge_df.drop("popularity" ,axis=1, inplace=True)

merge_df.drop("production_companies" ,axis=1, inplace=True)

merge_df.drop("release_date", axis=1, inplace = True)

merge_df.drop("production_countries", axis=1, inplace = True)

merge_df.drop("runtime", axis=1, inplace = True)

merge_df.drop("vote_average", axis=1, inplace = True)

merge_df.drop("vote_count", axis=1, inplace = True)

merge_df[merge_df["status"] == "Released"]

merge_df = merge_df[merge_df["status"] == "Released"]

print(type(merge_df["genres"].iloc[0]))

def clean_genre(x):
    if isinstance(x,str):
        try:
            return [i["name"] for i in ast.literal_eval(x)]
        except:
            return[]
    return[]

merge_df["genres"].apply(clean_genre)

merge_df["genres"] = merge_df["genres"].apply(clean_genre)

merge_df.head(2)

merge_df.join(merge_df["genres"].str.join("|").str.get_dummies())

merge_df = merge_df.join(merge_df["genres"].str.join("|").str.get_dummies())

merge_df.head(2)

# merge_df.drop("genres", axis=1, inplace=True)


print(merge_df['director'].unique())  # If encoding 'director'



def extract_director(crew_data):
    if isinstance(crew_data, str):  # Convert only if it's a string
        try:
            crew_data = ast.literal_eval(crew_data)  # Convert string to list
        except (ValueError, SyntaxError):
            return ''  # Return empty if conversion fails

    if isinstance(crew_data, list):  # Ensure it's now a list
        for i in crew_data:
            if i.get('job') == 'Director':  # Find director
                return i.get('name', '')  # Return director's name
    
    return ''  # Return empty if no director is found

# Apply function
# merge_df['director'] = merge_df['crew'].apply(extract_director)

merge_df = merge_df.join(pd.get_dummies(merge_df['director'], prefix='Director'))
# merge_df['director'] = merge_df['director'].replace('ABCD', '')  # Replace




merge_df = merge_df.join(merge_df['cast'].str.join('|').str.get_dummies())

merge_df.head(2)

merge_df['success_ratio'] = merge_df['revenue'] / merge_df['budget']


print(merge_df['success_ratio'].head())  # Correct way to get column data


merge_df.to_csv('/Users/ananyaaggarwal/Desktop/Movie Prediction Project/data/cleaned_and_merged_data.csv', index=False)
