==> notebooks/4_feature_engineering.ipynb <==
import seaborn as sns
import matplotlib.pyplot as plt
import pandas as pd
import numpy as np

merged_df = pd.read_csv('/Users/ananyaaggarwal/Desktop/Movie Prediction Project/data/cleaned_and_merged_data.csv')
# Feature engineering: Director impact
director_avg_revenue = merged_df.groupby('director')['revenue'].mean().reset_index()
director_avg_revenue.rename(columns={'revenue': 'director_avg_revenue'}, inplace=True)
merged_df = merged_df.merge(director_avg_revenue, on='director', how='left')

# Feature engineering: Cast impact
cast_avg_revenue = merged_df['cast'].explode().reset_index().merge(merged_df[['revenue']], left_on='index', right_index=True)
cast_avg_revenue = cast_avg_revenue.groupby('cast')['revenue'].mean().reset_index()
cast_avg_revenue.rename(columns={'revenue': 'cast_avg_revenue'}, inplace=True)
merged_df = merged_df.merge(cast_avg_revenue, left_on='cast', right_on='cast', how='left')

# Save the final processed dataset
merged_df.to_csv('/Users/ananyaaggarwal/Desktop/Movie Prediction Project/data/final_processed_data.csv', index=False)

merged_df.columns("director_avg_revenue")

print(director_avg_revenue.head())  # Ensure the director revenue data is calculated
print(director_avg_revenue.columns)  # Should contain 'director' and 'director_avg_revenue'


print(cast_avg_revenue.head())  # Ensure the cast revenue data is created
print(cast_avg_revenue.columns)  # Should contain 'cast' and 'cast_avg_revenue'
